requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pytest
httpx
pytest-asyncio
python-jose
supabase
